Refactor modification tests
